examples/community/requirements_wan_outpaint.txt
# Requirements for WAN Video Outpaint Examples
# Install with: pip install -r requirements_wan_outpaint.txt

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0

# Image and video processing
Pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Optional: for better performance
xformers>=0.0.20
safetensors>=0.3.0

# Optional: for video export
imageio>=2.31.0
imageio-ffmpeg>=0.4.8